sync data

+++ src/main/java/com/chiefmech/arms/entity/report/ShopDailyReport.java
@@ -4,11 +4,10 @@ public class ShopDailyReport {
 
 	private int dailyIncomingCount; // 每日入厂台数
 	private int monthlyIncomingCount;
-	private int dailyOutputValue; // 每日产值
-	private int monthlyOutputValue;
-	private int dailyGrossProfit; // 每日毛利
-	private int monthlyGrossProfit;
-
+	private float dailyOutputValue; // 每日产值
+	private float monthlyOutputValue;
+	private float dailyGrossProfit; // 每日毛利
+	private float monthlyGrossProfit;
 	public int getDailyIncomingCount() {
 		return dailyIncomingCount;
 	}
@@ -21,28 +20,28 @@ public class ShopDailyReport {
 	public void setMonthlyIncomingCount(int monthlyIncomingCount) {
 		this.monthlyIncomingCount = monthlyIncomingCount;
 	}
-	public int getDailyOutputValue() {
+	public float getDailyOutputValue() {
 		return dailyOutputValue;
 	}
-	public void setDailyOutputValue(int dailyOutputValue) {
+	public void setDailyOutputValue(float dailyOutputValue) {
 		this.dailyOutputValue = dailyOutputValue;
 	}
-	public int getMonthlyOutputValue() {
+	public float getMonthlyOutputValue() {
 		return monthlyOutputValue;
 	}
-	public void setMonthlyOutputValue(int monthlyOutputValue) {
+	public void setMonthlyOutputValue(float monthlyOutputValue) {
 		this.monthlyOutputValue = monthlyOutputValue;
 	}
-	public int getDailyGrossProfit() {
+	public float getDailyGrossProfit() {
 		return dailyGrossProfit;
 	}
-	public void setDailyGrossProfit(int dailyGrossProfit) {
+	public void setDailyGrossProfit(float dailyGrossProfit) {
 		this.dailyGrossProfit = dailyGrossProfit;
 	}
-	public int getMonthlyGrossProfit() {
+	public float getMonthlyGrossProfit() {
 		return monthlyGrossProfit;
 	}
-	public void setMonthlyGrossProfit(int monthlyGrossProfit) {
+	public void setMonthlyGrossProfit(float monthlyGrossProfit) {
 		this.monthlyGrossProfit = monthlyGrossProfit;
 	}
